Stock prices for the "SHANK" stocks.

A dataset containing the daily historical stock prices for the "SHANK" big korea stocks, "SAMSUNG", "HYUNDAI", "AMOREPACIFIC", "NAVER" and "KAKAO", spanning from 2012-02-08 through 2017-09-07.

SHANK

Format

A "tibble" ("tidy" data frame) with 7,580 rows and 8 variables:

symbol

stock ticker symbol

date

trade date

open

stock price at the open of trading, in won

high

stock price at the highest point during trading, in won

low

stock price at the lowest point during trading, in won

close

stock price at the close of trading, in won

volume

number of shares traded

adjusted

stock price at the close of trading adjusted for stock splits, in won