Get quantitative data from korea

Get quantitative data from korea

Usage

tqk_get(x, from = "1900-01-01", to = Sys.Date())

Arguments

x

Stock x. only Korean type like "005930" is samsung.

from

Optional for various time series functions. A character string representing a start date in YYYY-MM-DD format.

to

Optional for various time series functions. A character string representing a end date in YYYY-MM-DD format.

Value

a tibble

Examples

# \donttest{
  tqk_get(x = "005930")
#> # A tibble: 7,165 × 6
#>    date        open  high   low close   volume
#>    <date>     <int> <int> <int> <int>    <int>
#>  1 2023-08-29 66900 67200 66600 66800  9103907
#>  2 2023-08-28 66800 67000 66500 66800  5824628
#>  3 2023-08-25 67100 67400 66900 67100  7032462
#>  4 2023-08-24 68300 68700 67900 68200 15044463
#>  5 2023-08-23 66700 67100 66400 67100  9549352
#>  6 2023-08-22 67200 67700 66300 66600 10500242
#>  7 2023-08-21 66600 67100 66300 66600  9720067
#>  8 2023-08-18 66000 66700 65800 66300 11745006
#>  9 2023-08-17 66300 66800 66000 66700 10778652
#> 10 2023-08-16 66700 67100 66300 67000 13174578
#> # ℹ 7,155 more rows
  tqk_get(x = "005930", from = "2018-05-01")
#> # A tibble: 1,314 × 6
#>    date        open  high   low close   volume
#>    <date>     <int> <int> <int> <int>    <int>
#>  1 2023-08-29 66900 67200 66600 66800  9103907
#>  2 2023-08-28 66800 67000 66500 66800  5824628
#>  3 2023-08-25 67100 67400 66900 67100  7032462
#>  4 2023-08-24 68300 68700 67900 68200 15044463
#>  5 2023-08-23 66700 67100 66400 67100  9549352
#>  6 2023-08-22 67200 67700 66300 66600 10500242
#>  7 2023-08-21 66600 67100 66300 66600  9720067
#>  8 2023-08-18 66000 66700 65800 66300 11745006
#>  9 2023-08-17 66300 66800 66000 66700 10778652
#> 10 2023-08-16 66700 67100 66300 67000 13174578
#> # ℹ 1,304 more rows
  tqk_get(x = "005930", from = "2018-05-01", to = "2018-05-31")
#> # A tibble: 20 × 6
#>    date        open  high   low   close   volume
#>    <date>     <int> <int> <int>   <int>    <int>
#>  1 2018-05-31 50400 50800 49850   50700 63491109
#>  2 2018-05-30 51300 51500 49100   49500 20498098
#>  3 2018-05-29 52200 52500 51300   51300  8480437
#>  4 2018-05-28 52500 53000 52000   52300  9787820
#>  5 2018-05-25 51000 52800 50800   52700 15207266
#>  6 2018-05-24 52000 52000 51100   51400  8289275
#>  7 2018-05-23 50600 52000 50400   51800 17095490
#>  8 2018-05-21 49650 50200 49100   50000  9020998
#>  9 2018-05-18 49900 49900 49350   49500  6706570
#> 10 2018-05-17 50300 50500 49400   49400 10365440
#> 11 2018-05-16 49200 50200 49150   49850 15918683
#> 12 2018-05-15 50200 50400 49100   49200 18709146
#> 13 2018-05-14 51000 51100 49900   50100 14909272
#> 14 2018-05-11 52000 52200 51200   51300 10314997
#> 15 2018-05-10 51700 51700 50600   51600 13905263
#> 16 2018-05-09 52600 52800 50900   50900 16128305
#> 17 2018-05-08 52600 53200 51900   52600 23104720
#> 18 2018-05-04 53000 53900 51800   51900 39565391
#> 19 2018-05-03     0     0     0 2650000        0
#> 20 2018-05-02     0     0     0 2650000        0
# }